requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pyTelegramBotAPI==4.0.1
python-dotenv==0.19.0
requests==2.26.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
telebot==0.0.4
urllib3==1.26.6
Werkzeug==2.0.1
